Clarify documentation for `enable` pseudo-variable
The documentation for variables says that for boolean variables the form
`name1+name2+name3:var` is equivalent to
`name1:var & name2:var & name3:var` but it's possible to write
`name1+name2:enable` despite `enable` being the string "enable" or
"disable". This change clarifies that `enable` isn't really a variable
but syntactic sugar instead.